proto: revert UTF-8 validation for proto2
The proto specification officially says that proto2 and proto3 strings should be
validated, but pragmatically, compliance with the spec has been poor.
For example, the Go implementation did not validate either and added strict
validation recently to be compliant. However, this caused signficant breakage.

Cases of breakage should change the proto field type from string to the bytes type.
However, this is not always possible, when the field is part of the exposed API.
This tends to be the case for proto2, where some other notable language
implementations (like C++) do not validate proto2 for valid UTF-8.
However, since most language implementations do validate for UTF-8 in proto3,
we keep that behavior.

Making this change for Go is a little tricky since each field does not necessarily
know whether it is operating under the proto2 or proto3 syntax. Thus, we modify
the generator to emit a "utf8" struct field tag for any string fields in proto3.
The implications of this change is that people will need to regenerate their
proto files to have UTF-8 validation.

For completeness, the proto package checks for the utf8 tag in all cases even
if the field is obviously proto2 (i.e., is of the *string type).
However, the generator does not emit the utf8 tag for proto2, so this works fine.
We add support in the proto package because it is more consistent and also
for the event that we ever try to add validation to proto2 again.

We expand UTF-8 validation tests to ensure this works for the cross-product of
(proto2, proto3) and (scalar, vector, oneof, and maps) fields with strings.

Fixes #622